backend/requirements.txt
# Python dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.3.2
duckdb==0.9.2
pandas==2.1.4
openpyxl==3.1.2
redis==5.0.1
supabase==2.3.0
numpy==1.26.2